+++ src/frontend/src/app/create-schedule/create-schedule.component.css
@@ -1,3 +1,7 @@
+.card-header {
+    background-color: whitesmoke;
+}
+
 #rday {
     float: left;
     width: 30%;
