+++ src/frontend/src/app/scheduled-process/scheduled-process.component.html
@@ -39,14 +39,14 @@
           <app-update-modal (notify)='onNotify()' [childData]='updateScheduleData'></app-update-modal>
         </div>
       </div>
-      <div class="tile">
-        <div class="tile is-parent">
-          <div class="tile is-child card">
-            <app-create-schedule></app-create-schedule>
-          </div>
-        </div>
+    </div>
+    <div class="tile">
+    <div class="tile is-parent">
+      <div class="tile is-child">
+        <app-create-schedule></app-create-schedule>
       </div>
     </div>
+  </div>
   </div>
 
   <div class="tile">
