+++ src/frontend/src/app/scheduled-process/scheduled-process.component.ts
@@ -26,14 +26,8 @@ export class ScheduledProcessComponent implements OnInit {
   public tasksData = [];
   public days = [];
 
-  // To handle field validtion on UI
-  public invalidRepeat: boolean = false;
-  public invalidTime: boolean = false;
   public selectedTaskType = 'Latest'; // Default is LATEST
   public scheduler_name: string;
-
-  // Default selected schedule type is STARTUP = 1
-  public selected_schedule_type: Number = 1;
 
   // Object to hold schedule id to delete
   public childData: any;
@@ -81,8 +75,8 @@ export class ScheduledProcessComponent implements OnInit {
         }
         this.scheduleData = data.schedules;
         this.scheduleData.forEach(element => {
-          let repeatTimeObj = Utils.secondsToDhms(element.repeat);
-          if (repeatTimeObj.days == 1) {
+          const repeatTimeObj = Utils.secondsToDhms(element.repeat);
+          if (repeatTimeObj.days === 1) {
             element.repeat = repeatTimeObj.days + ' day, ' + repeatTimeObj.time;
           } else if (repeatTimeObj.days > 1) {
             element.repeat = repeatTimeObj.days + ' days, ' + repeatTimeObj.time;
@@ -117,7 +111,7 @@ export class ScheduledProcessComponent implements OnInit {
    * @param state Task state
    */
   public getTasks(state) {
-    if (state.toUpperCase() == 'RUNNING') {
+    if (state.toUpperCase() === 'RUNNING') {
       this.selectedTaskType = 'Running';
       this.getRunningTasks();
       return;
@@ -174,97 +168,13 @@ export class ScheduledProcessComponent implements OnInit {
           // TODO: remove cancelled task object from local list
           setTimeout (() => {
             console.log('waiting...', this.selectedTaskType);
-            if (this.selectedTaskType == 'Running') {
+            if (this.selectedTaskType === 'Running') {
               this.getRunningTasks();
             } else {
               this.getLatestTasks();
             }
           }, 5000);
         }
-      },
-      error => { console.log('error', error); });
-  }
-
-  /**
-   *  To set schedule type key globally for required field handling on UI
-   * @param value
-   */
-  public setScheduleTypeKey(value) {
-    this.selected_schedule_type = value;
-  }
-
-  public createSchedule() {
-    let schedule_name_fld = <HTMLInputElement>document.getElementById('name');
-    let schedule_process_fld = <HTMLSelectElement>document.getElementById('process');
-    let schedule_type_fld = <HTMLSelectElement>document.getElementById('type');
-
-    let exclusive_state_fld = <HTMLInputElement>document.getElementById('exclusive');
-
-    let repeat_day_fld = <HTMLInputElement>document.getElementById('rday');
-    let repeat_time_fld = <HTMLInputElement>document.getElementById('rtime');
-
-    // check if time is in valid range
-    this.invalidRepeat = Utils.not_between(repeat_time_fld.value);
-    if (this.invalidRepeat) {
-      return;
-    }
-
-    let day_fld = <HTMLSelectElement>document.getElementById('day');
-
-    let time_fld = <HTMLInputElement>document.getElementById('time');
-    let day = 0;
-    let scheduled_time = 0;
-
-
-    // total time with days and hh:mm:ss
-    let total_repeat_time = repeat_time_fld.value != '' ? Utils.convertTimeToSec(
-      repeat_time_fld.value, Number(repeat_day_fld.value)) : undefined;
-
-    /**
-     *  "schedule_type": [
-     *      {"index": 1, "name": "STARTUP"},
-     *      {"index": 2,"name": "TIMED"},
-     *      {"index": 3,"name": "INTERVAL"},
-     *      {"index": 4,"name": "MANUAL"}]
-     *      For schedule type 'TIMED', show 'Day' and 'TIME' field on UI
-     */
-    if (this.selected_schedule_type == 2) {  // Condition to check if schedule type is TIMED == 2
-      day = Number(day_fld.value);
-      // check if time is in valid range
-      this.invalidTime = Utils.not_between(time_fld.value);
-      if (this.invalidTime) {
-        return;
-      }
-      // Time value in seconds for TIMED schedule
-      scheduled_time = time_fld.value != '' ? Utils.convertTimeToSec(time_fld.value) : undefined;
-    }
-
-    let payload = {
-      'name': schedule_name_fld.value,
-      'process_name': schedule_process_fld.value,
-      'type': schedule_type_fld.value,
-      'repeat': total_repeat_time,
-      'day': day,
-      'time': scheduled_time,
-      'exclusive': exclusive_state_fld.checked
-    };
-    this.schedulesService.createSchedule(payload).
-      subscribe(
-      data => {
-         if (data.error) {
-          console.log('error in response', data.error);
-          this.alertService.error(data.error.message);
-          return;
-        }
-        this.getSchedules();
-
-        // Clear form fields
-        schedule_name_fld.value = '';
-        repeat_day_fld.value = '';
-        repeat_time_fld.value = '';
-        schedule_process_fld.value = this.scheduleProcess[0]; // set process dropdown to 0 index value
-        schedule_type_fld.value = '1';
-        this.setScheduleTypeKey(1); // To set schedule type key globally for required field handling on UI
       },
       error => { console.log('error', error); });
   }
